+++ resources/bootstrap/profiles/persistence/sql/app/src/{{project.files-name}}/boundary/port/persistence.clj
@@ -7,3 +7,5 @@
 
 ;; (defprotocol SomeProtocolName
 ;;   (some-method-name [db logger other-method-params...]))
+(defprotocol Health
+  (get-status [spec opts]))